fix(leaf-precision): eliminate float narrowing/string round-trip and R-side CSV floor for regression leaf values
Fitted regression (squared_error) leaf values were silently truncated to
~6-7 significant figures via two independent, stacked causes:

1. C++: Model::Model() cast the leaf mean to float then serialized it via
   std::to_string(float), which formats with a fixed 6 decimal places (not
   significant figures) -- catastrophic for small-magnitude leaves (e.g.
   -0.058940614050488 came back as -0.058941).
2. R: treefarms()/auto_tune_*() built the fit's CSV payload via
   apply(data_df, 1L, function(r) paste(as.character(r), collapse=",")),
   which forces the whole row -- including the unrelated y column -- through
   as.matrix.data.frame's homogeneous-type coercion, downgrading to 7-sig-fig
   formatting whenever any column (e.g. a factor-encoded feature) was
   non-numeric.

Fix (Oracle-reviewed across two rounds -- see quality_reports/plans/
2026-09-09_optimaltrees-leaf-value-precision-complete-fix.md):

- Model now keeps a `double regression_prediction_value` member, serialized
  at full round-trip precision (17 sig digits); _to_json() reads it directly,
  removing an unguarded, locale-dependent std::stod that silently fell back
  to 0.0 on parse failure.
- regression_targets_/target_values widened float->double throughout the
  encoder/dataset; new safe_stod() added beside (not replacing) safe_stof();
  internal loss/bound accumulators in Dataset::summary()'s SQUARED_ERROR
  branch widened to double for consistency (must ship with the storage
  widening, not separately -- otherwise the equiv-points safeguard compares
  a double-precision bound against a still-float-noisy sse). Task/Optimizer/
  Rashomon bound types and their float epsilon comparisons are unchanged --
  verified a hard, intentional containment boundary (Dataset::summary()'s
  min_loss/max_loss/potential are float&), not a same-diff site.
- R-side CSV builder rewritten column-wise (sprintf("%.17g", .) for numeric
  columns) in treefarms.R and auto_tune.R.

r-reviewer (quality_reports/reviews/2026-09-10_optimaltrees-csv-precision-fix.md)
caught a real regression in the R-side rewrite before commit: is.numeric(TRUE)
is FALSE, so the column-wise dispatch sent logical feature/outcome columns down
the character branch ("TRUE"/"FALSE") where the old apply()/as.matrix() path
coerced them to a double matrix first ("1"/"0") -- silently corrupting a
logical y in auto_tune_optimaltrees() (no upstream y <- as.numeric(y)
normalization there) into an NA prediction. Fixed: logical columns route
through the numeric branch in both files, with a regression test. Also fixed
per the same review: unname() the do.call(paste, ...) call, since a feature
column named "sep" or "collapse" would otherwise collide with paste()'s own
formals. Deferred, tracked, not fixed this pass: extracting the now-twice-
duplicated builder into one shared/validated helper, an exact (non-tolerance)
round-trip test of the CSV payload itself, and several low-severity polish
items -- see the review's "Resolution" section.

Verified: devtools::test() 0 failures / 2125 pass (2 pre-existing, unrelated
R-version-mismatch warnings in test-persistence.R). devtools::check() shows
2 errors / 12 warnings / 8 notes, all confirmed pre-existing (one error
dates to the package's first two commits) and none touching any file this
fix modified. test-leaf-value-precision.R (10 assertions, tolerance 1e-10)
covers small-magnitude leaves, a factor-column-in-X case (the R-side floor),
a logical-column case, and depth-2 multi-leaf trees.

Bounded reproducibility check against doubletree's single_tree_corollaries
Regime A cell (same committed seed): tree structure (n_leaves_m0, recovered,
realized_partition) and point estimates (theta) are bit-identical pre/post-
fix; mu_leaf_gap collapsed from ~4e-7 (matching the committed RDS exactly)
to ~2e-16 (machine epsilon). Confirms no retroactive change to tree
selection -- a pure precision fix.

Not in scope, named and deferred: split-threshold values still round to 6
decimals internally (encoder.cpp's std::to_string(threshold), self-
consistent for training); C++17 std::to_chars migration; broader silent-
catch cleanup in gosdt.cpp/model_set.cpp/rcpp_treefarms.cpp.

+++ src/dataset.cpp
@@ -307,7 +307,7 @@ void Dataset::summary(Bitmask const & capture_set, float & info, float & potenti
             max_loss = 0.0f;
             return;
         }
-        float sum_y = 0.0f;
+        double sum_y = 0.0;
         for (unsigned int i = 0; i < height(); ++i) {
             if (capture_set.get(i)) {
                 sum_y += this -> target_values[i];
@@ -316,26 +316,31 @@ void Dataset::summary(Bitmask const & capture_set, float & info, float & potenti
         if (n == 0) {
             throw std::runtime_error("Cannot compute mean of empty sample");
         }
-        float mean_y = sum_y / (float)n;
+        double mean_y = sum_y / (double)n;
         if (!std::isfinite(mean_y)) {
             throw std::runtime_error("Non-finite mean: sum=" + std::to_string(sum_y) + ", n=" + std::to_string(n));
         }
-        float sse = 0.0f;
+        double sse = 0.0;
         for (unsigned int i = 0; i < height(); ++i) {
             if (capture_set.get(i)) {
-                float d = this -> target_values[i] - mean_y;
+                double d = this -> target_values[i] - mean_y;
                 sse += d * d;
             }
         }
-        min_loss = sse;
-        max_loss = sse;
+        // Narrow once here: Dataset::summary's min_loss/max_loss/potential are
+        // float& (an intentional, unwidened boundary -- see 2026-09-09 precision
+        // fix plan; Task/Optimizer/Rashomon bounds downstream all stay float and
+        // must not be perturbed by this fix). Everything above and below this
+        // point up to those three assignments accumulates in double.
+        min_loss = static_cast<float>(sse);
+        max_loss = static_cast<float>(sse);
 
         // Compute lower bounds for regression (n > 1 required for meaningful bounds)
         if (n > 1) {
             // ALWAYS compute equivalent points bound (OSRT approach when k_cluster=false)
             // Group samples by feature vector - samples with identical features
             // must receive the same prediction, so within-group variance is unavoidable
-            std::map<Bitmask, std::vector<float>> equiv_groups;
+            std::map<Bitmask, std::vector<double>> equiv_groups;
 
             for (unsigned int i = 0; i < height(); ++i) {
                 if (capture_set.get(i)) {
@@ -348,11 +353,11 @@ void Dataset::summary(Bitmask const & capture_set, float & info, float & potenti
             // This is the MINIMUM achievable loss because samples with same features
             // must get the same prediction, so variance within each group is unavoidable
             std::vector<double> weights, values;
-            float equiv_points_loss = 0.0f;  // Sum of within-group SSE
+            double equiv_points_loss = 0.0;  // Sum of within-group SSE
 
             // C++11 compatible iteration (no structured bindings)
             for (auto it = equiv_groups.begin(); it != equiv_groups.end(); ++it) {
-                const std::vector<float> & targets = it->second;
+                const std::vector<double> & targets = it->second;
                 double w = (double)targets.size();
                 double sum = 0.0;
                 double sum_sq = 0.0;
@@ -378,7 +383,7 @@ void Dataset::summary(Bitmask const & capture_set, float & info, float & potenti
             }
 
             // Use equiv_points_loss as base lower bound
-            float min_achievable_loss = equiv_points_loss;
+            double min_achievable_loss = equiv_points_loss;
 
             // If k-means enabled, try to get an even tighter bound
             if (Configuration::k_cluster) {
@@ -393,23 +398,24 @@ void Dataset::summary(Bitmask const & capture_set, float & info, float & potenti
                 // k-Means bound includes both:
                 // - Between-cluster SSE (from k-means on aggregated points)
                 // - Within-cluster SSE (from equiv_points_loss)
-                min_achievable_loss = (float)(kmeans_sse + equiv_points_loss);
+                min_achievable_loss = (double)(kmeans_sse + equiv_points_loss);
             }
 
             // Safeguard: bound should never exceed current SSE
-            if (min_achievable_loss > sse || min_achievable_loss < 0.0f) {
+            if (min_achievable_loss > sse || min_achievable_loss < 0.0) {
                 // Bound is invalid - fall back to equiv points only
                 min_achievable_loss = equiv_points_loss;
             }
 
             // Ensure bound is non-negative and doesn't exceed current loss
-            min_achievable_loss = std::max(0.0f, std::min(min_achievable_loss, sse));
+            min_achievable_loss = std::max(0.0, std::min(min_achievable_loss, sse));
 
-            // potential = maximum possible reduction in loss
-            potential = sse - min_achievable_loss;
+            // potential = maximum possible reduction in loss. Narrow once here,
+            // same boundary as min_loss/max_loss above.
+            potential = static_cast<float>(sse - min_achievable_loss);
 
             // Update min_loss to reflect lower bound
-            min_loss = min_achievable_loss;
+            min_loss = static_cast<float>(min_achievable_loss);
         } else {
             // n <= 1: no room for improvement (single sample or empty)
             potential = 0.0f;
