+++ src/encoder.hpp
@@ -68,7 +68,7 @@ public:
     unsigned int samples(void) const;
 
     // Regression (squared_error loss): continuous target values, one per sample. Empty when not regression.
-    std::vector< float > const & regression_targets(void) const;
+    std::vector< double > const & regression_targets(void) const;
 
     // Returns: the pre-encode feature index and the offset to the binary feature generated
     // @param encoded_column_index: the index of the binary feature
@@ -143,8 +143,10 @@ private:
     // Binary representation of rows
     std::vector< Bitmask > binary_rows;
 
-    // Regression mode: continuous target column stored as floats (when loss_function == SQUARED_ERROR)
-    std::vector< float > regression_targets_;
+    // Regression mode: continuous target column, parsed at full double precision
+    // via safe_stod (2026-09-09 fix; encoder.cpp's safe_stof stays float and is
+    // still used for feature-threshold parsing elsewhere in this file).
+    std::vector< double > regression_targets_;
 
     // @param number: input number to reduce precision
     // @return an input equivalent to number rounded to k significant figures
